+++ lua/quicktest/adapters/meson/util.lua
@@ -1,4 +1,3 @@
-local Job = require("plenary.job")
 local json5 = require("json5")
 local meson = require("quicktest.adapters.meson.meson")
 
@@ -48,17 +47,6 @@ function M.get_filename(path)
   return path:match("[^/]*.c$")
 end
 
-local function mysplit(inputstr, sep)
-  if sep == nil then
-    sep = "%s"
-  end
-  local t = {}
-  for str in string.gmatch(inputstr, "([^" .. sep .. "]+)") do
-    table.insert(t, str)
-  end
-  return t
-end
-
 -- Uses meson introspect CLI to find the name of the test
 -- executable from the path of the open file
 function M.get_test_exe_from_buffer(bufnr)
@@ -77,30 +65,6 @@ function M.get_test_exe_from_buffer(bufnr)
     end
   end
   return nil
-end
-
-function M.get_test_suite_and_name(line)
-  local parts = mysplit(line, "(")
-  local index = 1
-
-  if vim.startswith(parts[1], "ParameterizedTest") then
-    index = 2
-  end
-
-  parts = mysplit(parts[2], ",")
-  return {
-    test_suite = string.gsub(parts[index], "%s+", ""),
-    test_name = string.gsub(parts[index + 1], "%s+", ""),
-  }
-end
-
-function M.get_nearest_test(bufnr, cursor_pos)
-  for pos = cursor_pos[1], 1, -1 do
-    local line = vim.api.nvim_buf_get_lines(bufnr, pos - 1, pos, true)[1]
-    if vim.startswith(line, "Test(") or vim.startswith(line, "ParameterizedTest(") then
-      return line
-    end
-  end
 end
 
 function M.make_test_args(params)
